requirements.txt
# Core dependencies
python>=3.8
click>=8.0.0
pyyaml>=6.0
requests>=2.25.0
python-dotenv>=0.19.0
colorama>=0.4.4
jinja2>=3.0.0
python-dateutil>=2.8.0

# Cloud SDKs
google-cloud-run>=0.9.0
google-cloud-sql>=2.0.0
google-cloud-build>=3.0.0
google-cloud-artifact-registry>=1.0.0
google-api-python-client>=2.0.0

# Infrastructure
docker>=5.0.0
kubernetes>=20.0.0

# Utilities
progress>=1.6
rich>=10.0.0
tabulate>=0.8.0
pygments>=2.10.0

# Testing
pytest>=6.0.0
pytest-mock>=3.6.0
pytest-asyncio>=0.15.0
# Ollama Integration
aiohttp>=3.8.0